requirements.txt
streamlit
openai
google-generativeai
PyMuPDF